api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
psycopg2-binary
httpx
google-adk
litellm
google-generativeai
anthropic
resend
python-jose[cryptography]
